# src/main/java/com/fleet/maintenance/auth/controller/AuthController.java
package com.fleet.maintenance.auth.controller;

import com.fleet.maintenance.auth.dto.LoginRequest;
import com.fleet.maintenance.auth.dto.LoginResponse;
import com.fleet.maintenance.auth.service.AuthService;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class AuthController {
    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/auth/login")
    public ResponseEntity<LoginResponse> login(@Valid @RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }
}