# src/main/resources/maintanence-openapi.yaml
openapi: 3.0.3
info:
  title: Smart Fleet Maintenance and Service Authorization API
  version: 1.0.0
  description: >
    OpenAPI contract for a Spring Boot 3 and Java 21 smart fleet maintenance
    platform using modular domain-driven bounded contexts for request,
    inspection, and decision workflows.
servers:
  - url: http://localhost:8080/api
    description: Local dev

security:
  - bearerAuth: []

tags:
  - name: Auth
    description: Authentication and JWT issuance.
  - name: Dashboard
    description: Aggregated read projection across request, inspection, and decision contexts.
  - name: Request
    description: Maintenance request bounded context.
  - name: Inspection
    description: Inspection bounded context for provider reports and estimates.
  - name: Decision
    description: Service authorization decision bounded context.

paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate a coordinator or provider
      description: Public endpoint. Returns a JWT bearer token containing id, name, and role claims.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /dashboard:
    get:
      tags:
        - Dashboard
      summary: Get dashboard maintenance projection
      description: >
        Returns an aggregated projection from multiple domains. Results are
        filtered by coordinatorId=userId from JWT for coordinators, or
        assignedTo=userId from JWT for providers.
      operationId: getDashboard
      responses:
        '200':
          description: Dashboard entries visible to the authenticated user.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DashboardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /maintenance-requests:
    post:
      tags:
        - Request
      summary: Create a maintenance request
      description: Coordinator creates a new maintenance request.
      operationId: createMaintenanceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceRequest'
      responses:
        '201':
          description: Maintenance request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /maintenance-requests/{id}:
    get:
      tags:
        - Request
      summary: Get maintenance request details
      description: Returns detailed request information, including inspection and decision state when available.
      operationId: getMaintenanceRequestById
      parameters:
        - $ref: '#/components/parameters/MaintenanceIdPath'
      responses:
        '200':
          description: Maintenance request details.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/MaintenanceRequestDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /maintenance-requests/{id}/assign-provider:
    put:
      tags:
        - Request
      summary: Assign a provider to a maintenance request
      description: Coordinator assigns a provider and moves the request toward provider inspection.
      operationId: assignProvider
      parameters:
        - $ref: '#/components/parameters/MaintenanceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignProviderRequest'
      responses:
        '200':
          description: Provider assigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /maintenance-requests/{id}/inspection:
    post:
      tags:
        - Inspection
      summary: Submit inspection report and estimate
      description: Provider submits an inspection report, cost estimate, and expected completion date.
      operationId: submitInspection
      parameters:
        - $ref: '#/components/parameters/MaintenanceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionRequest'
      responses:
        '200':
          description: Inspection submitted and request is pending coordinator approval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /maintenance-requests/{id}/decision:
    post:
      tags:
        - Decision
      summary: Submit coordinator decision
      description: >
        Coordinator approves, rejects, or requests more information. On approval,
        the backend publishes a PAYMENT_INITIATED Kafka event; this event is not
        exposed as an API endpoint.
      operationId: submitDecision
      parameters:
        - $ref: '#/components/parameters/MaintenanceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: Decision submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    MaintenanceIdPath:
      name: id
      in: path
      required: true
      description: Maintenance request identifier.
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or JWT is missing/invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user does not have permission to perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested maintenance resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    LoginRequest:
      type: object
      description: Java 21 record candidate with Bean Validation annotations on required fields.
      additionalProperties: false
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          example: coordinator1
        password:
          type: string
          minLength: 1
          format: password
          example: password

    LoginResponse:
      type: object
      description: JWT login response. Token claims include id, name, and role.
      additionalProperties: false
      required:
        - username
        - token
      properties:
        username:
          type: string
          example: coordinator1
        token:
          type: string
          description: JWT bearer token.
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJjb29yZGluYXRvcjEifQ.signature

    DashboardResponse:
      type: object
      description: Aggregated dashboard projection from request, inspection, and decision bounded contexts.
      additionalProperties: false
      required:
        - totalCount
        - items
      properties:
        totalCount:
          type: integer
          description: Total number of dashboard items
        items:
          type: array
          description: All items array
          items:
            $ref: '#/components/schemas/DashboardItem'

    DashboardItem:
      type: object
      description: Aggregated dashboard projection from request, inspection, and decision bounded contexts.
      additionalProperties: false
      required:
        - request
        - inspection
      properties:
        request:
          $ref: '#/components/schemas/DashboardRequest'
        inspection:
          $ref: '#/components/schemas/DashboardInspection'
          nullable: true

    DashboardRequest:
      type: object
      required:
        - maintenanceId
        - vehicleId
        - status
        - description
        - createdOn
        - requesterName
      properties:
        maintenanceId:
          type: string
          format: uuid
        vehicleId:
          type: string
        status:
          $ref: '#/components/schemas/MaintenanceStatus'
        description:
          type: string
        createdOn:
          type: string
          format: date-time
        requesterName:
          type: string

    DashboardInspection:
      type: object
      required:
        - inspectionId
        - estimatedAmount
        - inspectionDate
      properties:
        inspectionId:
          type: string
          format: uuid
        estimatedAmount:
          type: number
          format: double
        inspectionDate:
          type: string
          format: date-time

    CreateMaintenanceRequest:
      type: object
      description: Request-context Java 21 record candidate for creating maintenance requests.
      additionalProperties: false
      required:
        - vehicleId
        - description
        - severity
        - impact
        - impactedPeopleCount
        - createdOn
      properties:
        vehicleId:
          type: string
          minLength: 1
          example: VH-1024
        description:
          type: string
          minLength: 1
          example: Engine warning light is active and vehicle has reduced power.
        severity:
          type: string
          minLength: 1
          example: HIGH
        impact:
          type: string
          minLength: 1
          example: Vehicle unavailable for delivery route.
        impactedPeopleCount:
          type: integer
          format: int32
          minimum: 0
          example: 3
        createdOn:
          type: string
          format: date-time

    AssignProviderRequest:
      type: object
      description: Request-context Java 21 record candidate for assigning a provider.
      additionalProperties: false
      required:
        - maintenanceId
        - providerName
        - providerId
        - updatedOn
      properties:
        maintenanceId:
          type: string
          format: uuid
        providerName:
          type: string
          minLength: 1
          example: Acme Fleet Service
        providerId:
          type: string
          minLength: 1
          example: provider-101
        updatedOn:
          type: string
          format: date-time

    InspectionRequest:
      type: object
      description: Inspection-context Java 21 record candidate for provider report and estimate submission.
      additionalProperties: false
      required:
        - maintenanceId
        - updatedOn
        - inspectionReport
        - estimatedCost
        - inspectedOn
        - estimatedCompletionDate
      properties:
        maintenanceId:
          type: string
          format: uuid
        updatedOn:
          type: string
          format: date-time
        inspectionReport:
          type: string
          minLength: 1
          example: Alternator failure confirmed. Replacement required.
        estimatedCost:
          type: number
          format: double
          minimum: 0
          example: 850.75
        inspectedOn:
          type: string
          format: date-time
        estimatedCompletionDate:
          type: string
          format: date-time
        additionalDetails:
          type: string
          example: Parts available within one business day.

    DecisionRequest:
      type: object
      description: Decision-context Java 21 record candidate for coordinator authorization decisions.
      additionalProperties: false
      required:
        - maintenanceId
        - decisionType
        - remarks
        - updatedOn
      properties:
        maintenanceId:
          type: string
          format: uuid
        decisionType:
          $ref: '#/components/schemas/DecisionType'
        remarks:
          type: string
          minLength: 1
          example: Approved for immediate repair.
        updatedOn:
          type: string
          format: date-time

    MaintenanceStatusResponse:
      type: object
      description: Common status response returned by workflow commands.
      additionalProperties: false
      required:
        - maintenanceId
        - status
      properties:
        maintenanceId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MaintenanceStatus'

    MaintenanceRequestDetails:
      type: object
      description: Detailed maintenance request view assembled from request, inspection, and decision contexts.
      additionalProperties: false
      required:
        - maintenanceId
        - status
        - vehicleId
        - requesterId
        - description
        - severity
        - impact
        - impactedPeopleCount
        - createdOn
        - updatedOn
      properties:
        maintenanceId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MaintenanceStatus'
        vehicleId:
          type: string
          minLength: 1
        requesterId:
          type: string
          minLength: 1
        description:
          type: string
        severity:
          type: string
        impact:
          type: string
        impactedPeopleCount:
          type: integer
          format: int32
          minimum: 0
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
        assignedTo:
          type: string
        estimatedCost:
          type: number
          format: double
          minimum: 0
        inspectionReport:
          type: string
        inspectedOn:
          type: string
          format: date-time
        estimatedCompletionDate:
          type: string
          format: date-time
        additionalDetails:
          type: string

    MaintenanceStatus:
      type: string
      enum:
        - CREATED
        - ASSIGNED
        - PENDING_APPROVAL
        - APPROVED
        - REJECTED
        - RFI_REQUESTED
        - PAYMENT_INITIATED

    DecisionType:
      type: string
      enum:
        - APPROVE
        - REJECT
        - REQUEST_MORE_INFO

    Role:
      type: string
      description: Application role supplied in JWT claims.
      enum:
        - ROLE_COORDINATOR
        - ROLE_PROVIDER

    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Invalid username or password.
        path:
          type: string
          example: /api/auth/login